implement some custom status codes

+++ src/main/java/lv/learning/spring_boot_with_mysql/service/EmployeeService.java
@@ -33,52 +33,62 @@ public class EmployeeService {
 
     // read a single Employee record
     public EmployeeRest readEmployee(Integer emp_no) {
-        EmployeeRest employeeRest = employeeRepository.findById(emp_no).orElse(null);
-        return employeeRest;
+//        EmployeeRest storedEmployee = employeeRepository.findById(emp_no).orElse(null);
+//        return storedEmployee;
+        return employeeRepository.findById(emp_no).orElse(null);
     }
 
     // create new Employee record
     public EmployeeRest createEmployee(EmployeeRequest employeeRequest) {
-        System.out.println("Gender char before = " + employeeRequest.getGender());
+//        System.out.println("Gender char before = " + employeeRequest.getGender());
         employeeRequest.setGender(Character.toUpperCase(employeeRequest.getGender()));
-        System.out.println("Gender char after = " + employeeRequest.getGender());
-        EmployeeRest newEmployeeRest = new EmployeeRest();
-        newEmployeeRest.setBirth_date(employeeRequest.getBirth_date());
-        newEmployeeRest.setFirst_name(employeeRequest.getFirst_name());
-        newEmployeeRest.setLast_name(employeeRequest.getLast_name());
-        newEmployeeRest.setGender(employeeRequest.getGender());
-        newEmployeeRest.setHire_date(employeeRequest.getHire_date());
-        return employeeRepository.save(newEmployeeRest);
+//        System.out.println("Gender char after = " + employeeRequest.getGender());
+        EmployeeRest newEmployee = new EmployeeRest();
+        newEmployee.setBirth_date(employeeRequest.getBirth_date());
+        newEmployee.setFirst_name(employeeRequest.getFirst_name());
+        newEmployee.setLast_name(employeeRequest.getLast_name());
+        newEmployee.setGender(employeeRequest.getGender());
+        newEmployee.setHire_date(employeeRequest.getHire_date());
+        return employeeRepository.save(newEmployee);
     }
 
     // update an Employee record
-    public EmployeeRest updateEmployee(Integer emp_no, EmployeeRequest employeeRequest) {
-        EmployeeRest existingEmployeeRest = employeeRepository.findById(emp_no).orElse(null);
-        if (existingEmployeeRest != null) {
-            if (employeeRequest.getBirth_date() != null && !(employeeRequest.getBirth_date().equalsIgnoreCase(existingEmployeeRest.getBirth_date()))) {
-                existingEmployeeRest.setBirth_date(employeeRequest.getBirth_date());
+    public EmployeeRest updateEmployee(Integer emp_no, EmployeeRequest updateRequest) {
+        EmployeeRest storedEmployee = employeeRepository.findById(emp_no).orElse(null);
+        if (storedEmployee != null) {
+            if (updateRequest.getBirth_date() != null && !(updateRequest.getBirth_date().equalsIgnoreCase(storedEmployee.getBirth_date()))) {
+                storedEmployee.setBirth_date(updateRequest.getBirth_date());
             }
-            if (employeeRequest.getFirst_name() != null && !(employeeRequest.getFirst_name().equalsIgnoreCase(existingEmployeeRest.getFirst_name()))) {
-                existingEmployeeRest.setFirst_name(employeeRequest.getFirst_name());
+            if (updateRequest.getFirst_name() != null && !(updateRequest.getFirst_name().equalsIgnoreCase(storedEmployee.getFirst_name()))) {
+                storedEmployee.setFirst_name(updateRequest.getFirst_name());
             }
-            if (employeeRequest.getLast_name() != null && !(employeeRequest.getLast_name().equalsIgnoreCase(existingEmployeeRest.getLast_name()))) {
-                existingEmployeeRest.setLast_name(employeeRequest.getLast_name());
+            if (updateRequest.getLast_name() != null && !(updateRequest.getLast_name().equalsIgnoreCase(storedEmployee.getLast_name()))) {
+                storedEmployee.setLast_name(updateRequest.getLast_name());
             }
-            if (employeeRequest.getGender() != null) {
-                employeeRequest.setGender(Character.toUpperCase(employeeRequest.getGender()));
-                if (!(employeeRequest.getGender().equals(existingEmployeeRest.getGender()))) {
-                    existingEmployeeRest.setGender(employeeRequest.getGender());
+            if (updateRequest.getGender() != null) {
+                updateRequest.setGender(Character.toUpperCase(updateRequest.getGender()));
+                if (!(updateRequest.getGender().equals(storedEmployee.getGender()))) {
+                    storedEmployee.setGender(updateRequest.getGender());
                 }
             }
-            if (employeeRequest.getHire_date() != null && !(employeeRequest.getHire_date().equalsIgnoreCase(existingEmployeeRest.getHire_date()))) {
-                existingEmployeeRest.setHire_date(employeeRequest.getHire_date());
+            if (updateRequest.getHire_date() != null && !(updateRequest.getHire_date().equalsIgnoreCase(storedEmployee.getHire_date()))) {
+                storedEmployee.setHire_date(updateRequest.getHire_date());
             }
+
+            return employeeRepository.save(storedEmployee);
+        } else {
+            return null;
         }
-        return employeeRepository.save(existingEmployeeRest);
     }
 
     // delete an Employee record
-    public void deleteEmployee(Integer emp_no) {
-        employeeRepository.deleteById(emp_no);
+    public short deleteEmployee(Integer emp_no) {
+        if (employeeRepository.findById(emp_no).orElse(null) != null) {
+            employeeRepository.deleteById(emp_no);
+            return 204;
+        }
+        else {
+            return 404;
+        }
     }
 }
